requirements.txt
streamlit==1.36.0
pyairtable==2.3.3
relevanceai==10.0.4
python-dotenv==1.0.1
